Data_Generation/requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.0
langchain-core==0.3.22
langsmith==0.1.147
orjson==3.10.12
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
typing_extensions==4.12.2
urllib3==2.2.3
Django==5.1
openai
markdown
pdfkit
langchain-openai
ragas
unstructured
llama-index
